requirements.txt
flask
pillow
numpy
python-magic; platform_system != "Windows"
python-magic-bin; platform_system == "Windows"
PyPDF2
scipy
pydub
matplotlib